requirements.txt
Django~=4.1.4
djangorestframework==3.14
python-dotenv~=0.21.0
django-autoslug~=1.9.8